feat: start external controller alongside existing app bootstrap
